Gather common tests
