+++ Tests/PlayerTests/PlayerTests.swift
@@ -12,14 +12,6 @@ import XCTest
 
 @MainActor
 final class PlayerTests: XCTestCase {
-    func testEmpty() {
-        let player = Player()
-        expect(player.currentItem).to(beNil())
-        expect(player.items).to(beEmpty())
-        expect(player.nextItems).to(beEmpty())
-        expect(player.previousItems).to(beEmpty())
-    }
-
     func testPlayerDeallocation() {
         let item = AVPlayerItem(url: Stream.onDemand.url)
         var player: Player? = Player(item: item)
